Separate the decorator into it’s own file Make the decorator raise an exception if not wrapping a django class-based view Updated tests and README
