+++ thumber/__init__.py
@@ -2,5 +2,5 @@ default_app_config = 'thumber.apps.ThumberConfig'
 
 
 def _expose_items():
-    from .views import thumber_feedback
+    from .decorators import thumber_feedback
     globals()['thumber_feedback'] = thumber_feedback
